Refactor IdempotencyPropertyTest into CorrectnessTest and PropertyTest
PropertyTest contains all the logic to drive the property test.
* checks for regression
* progressbar
* todo file for failures

CorrectnessTest contains the meat for the property test.

CorrectnessTest
